+++ src/nylas-connection.js
@@ -129,6 +129,10 @@ module.exports = class NylasConnection {
 
     return new Promise((resolve, reject) => {
       return request(options, (error, response, body = {}) => {
+        if (typeof response === 'undefined') {
+          error = new Error('Response is undefined');
+          return reject(error);
+        }
         // node headers are lowercase so this refers to `Nylas-Api-Version`
         const apiVersion = response.headers['nylas-api-version'];